requirements_web.txt
Flask
mysql-connector-python
pandas
numpy
beautifulsoup4
googlesearch-python
requests
selenium
webdriver-manager
google-generativeai
httpx
